queue: radically change queueing system
This is slightly nicer to use, maybe, but is way more complicated and
I guess for no good reason. I reckon easier would be some kind of
dummy Garmin object that sat around when there was no connection. This
would all be saved if I fixed the main problem in that Garmin objects
can't be re-used after they've been disconnected.