activities: ensure start time is a string when saved to config
ConfigParser will save values to disk serialized but if the
ConfigParser object remains alive, a subsequent .get() call will
return the object.

In this case, a datetime object was being saved. On the first run we
got a string because that's the serialized format, but after
disconnecting, and reconnecting (but keeping the same ConfigParser
around), we got a datetime instead which couldn't be parsed as a
string or even interpolated by ConfigParser.

This patch adds some extra serialization and date parsing but it makes
the code simpler.